requirements.txt
# SAP CPI Self-Healing Agent — production dependencies
# Generated from pyproject.toml direct dependencies.
# playwright is excluded (browser binaries unavailable in CF; used only in scrape_sap_docs.py utility).
# asyncio is excluded (stdlib since Python 3.4).

aiolimiter>=1.2.1
duckduckgo-search>=7.0.0
anyio>=4.9.0
beautifulsoup4>=4.12.0
boto3>=1.34.0
cachetools>=5.0.0
fastapi>=0.115.0
fastmcp>=2.14.5
hdbcli>=2.21.0
httpx>=0.27.0
langchain>=0.3.0
langchain-mcp-adapters>=0.2.1
prometheus-client>=0.20.0
pydantic>=2.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
reportlab>=4.0.0
sap-ai-sdk-gen>=5.10.0
structlog>=24.0.0
tabulate>=0.9.0
uvicorn[standard]>=0.29.0
yaspin>=3.0.0
